Move BackendName to core/base/common
