+++ unidist/cli/__main__.py
@@ -15,12 +15,13 @@ except ImportError:
         0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../.."))
     )
 
-from unidist.cli.base.utils import BackendName, Defaults
+from unidist.cli.base.utils import Defaults
 from unidist.cli.backends.ray.runner import RayRunner
 from unidist.cli.backends.mpi.runner import MPIRunner
 from unidist.cli.backends.dask.runner import DaskRunner
 from unidist.cli.backends.python.runner import PythonRunner
 from unidist.cli.backends.multiprocessing.runner import MultiProcessingRunner
+from unidist.core.base.common import BackendName
 
 
 def main():
